Refactor source code + test for Full matrix + delete old tests
