+++ src/matrix_lib/__init__.py
@@ -1,9 +1,11 @@
-from .base_matrix import MatrixBase
-from .matrix import FullMatrix, Matrix, SymmetricMatrix
+from .full_matrix import FullMatrix
+from .matrix import Matrix
+from .symmetric_matrix import SymmetricMatrix
+from .triangular_matrix import TriangularMatrix
 
 
 def hello() -> str:
     return "Hello from matrix-lib!"
 
 
-__all__ = ["MatrixBase", "Matrix", "FullMatrix", "SymmetricMatrix"]
+__all__ = ["MatrixBase", "Matrix", "FullMatrix", "SymmetricMatrix", "TriangularMatrix"]
