+++ src/matrix_lib/full_matrix.py
@@ -1,0 +1,50 @@
+import numpy as np
+
+from .matrix import Matrix
+
+
+class FullMatrix(Matrix):
+    """
+    Заполненная матрица с элементами произвольного типа.
+    """
+
+    def __init__(self, data):
+        """
+        Создает объект, хранящий матрицу в виде np.ndarray `data`.
+        """
+        assert isinstance(data, np.ndarray)
+        self.data = data
+
+    def empty_like(self, width=None, height=None):
+        dtype = self.data.dtype
+        if width is None:
+            width = self.data.shape[1]
+        if height is None:
+            height = self.data.shape[0]
+        data = np.empty((height, width), dtype=dtype)
+        return FullMatrix(data)
+
+    @classmethod
+    def zero(_cls, height, width, default=0):
+        """
+        Создает матрицу размера `width` x `height` со значениями по умолчанию `default`.
+        """
+        data = np.empty((height, width), dtype=type(default))
+        data[:] = default
+        return FullMatrix(data)
+
+    @property
+    def shape(self):
+        return self.data.shape
+
+    @property
+    def dtype(self):
+        return self.data.dtype
+
+    def __getitem__(self, key):
+        row, column = key
+        return self.data[row, column]
+
+    def __setitem__(self, key, value):
+        row, column = key
+        self.data[row, column] = value
